refactor backfill to background job
